Refactoring and extended request history by 10
